Rename Objects folder and namespace.
