+++ SubnauticaRandomiser/Logic/CoreLogic.cs
@@ -5,8 +5,8 @@ using System.Linq;
 using JetBrains.Annotations;
 using SubnauticaRandomiser.Interfaces;
 using SubnauticaRandomiser.Logic.Recipes;
-using SubnauticaRandomiser.RandomiserObjects;
-using SubnauticaRandomiser.RandomiserObjects.Enums;
+using SubnauticaRandomiser.Objects;
+using SubnauticaRandomiser.Objects.Enums;
 
 namespace SubnauticaRandomiser.Logic
 {
