+++ SubnauticaRandomiser/Logic/ProgressionTree.cs
@@ -2,8 +2,8 @@
 using System.Linq;
 using JetBrains.Annotations;
 using SubnauticaRandomiser.Logic.Recipes;
-using SubnauticaRandomiser.RandomiserObjects;
-using SubnauticaRandomiser.RandomiserObjects.Enums;
+using SubnauticaRandomiser.Objects;
+using SubnauticaRandomiser.Objects.Enums;
 
 namespace SubnauticaRandomiser.Logic
 {
