+++ SubnauticaRandomiser/Logic/Recipes/ModeRandom.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
 using SMLHelper.V2.Handlers;
 using SubnauticaRandomiser.Interfaces;
-using SubnauticaRandomiser.RandomiserObjects;
-using SubnauticaRandomiser.RandomiserObjects.Enums;
+using SubnauticaRandomiser.Objects;
+using SubnauticaRandomiser.Objects.Enums;
 
 namespace SubnauticaRandomiser.Logic.Recipes
 {
