+++ Tests/UnitTests/Logic/Recipes/MaterialsTest.cs
@@ -1,8 +1,8 @@
 using System.Collections.Generic;
 using NUnit.Framework;
 using SubnauticaRandomiser.Logic.Recipes;
-using SubnauticaRandomiser.RandomiserObjects;
-using SubnauticaRandomiser.RandomiserObjects.Enums;
+using SubnauticaRandomiser.Objects;
+using SubnauticaRandomiser.Objects.Enums;
 using Tests.Mocks;
 
 namespace Tests.UnitTests.Logic.Recipes
